Refactor searchBacklog -> search/backlog
